Bugfix: don't allow shks and priests to walk onto the player's square
mfndpos returns the square that the monster wants to move to. If
ALLOW_MUXY is unset, it'll never move to the square it thinks you're
on. However, if you're invisible/displaced/etc., it can sometimes
select the square you're actually on (because it thinks you are
somewhere else and thus that square is free); we need to make sure
that such a move can't be successful, because the player and
monster can't occupy the same square.