Allow light sources to be recalculated rather than dead-reckoned
This fixes a bug involving migrating monsters wielding Sunsword:
we suppress the light source while the monster is on the migration
chain, then recalculate it when the monster arrives. This approach
also generally makes the light source code cleaner (it could be
made cleaner still but not all the code is translated over to the
new approach).

Unfortunately, we probably can't use the same approach to simplify
the migration of timers...